requirements.txt
# Khet Guard ML Inference API Requirements
# Core ML and API dependencies
fastapi==0.104.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7

# ML and Computer Vision
torch==2.8.0
torchvision==0.23.0
onnxruntime==1.22.1
onnx==1.19.0
numpy==1.26.4
pillow==10.1.0

# Visualization and Analysis
matplotlib==3.10.6

# Environment and Configuration
python-dotenv==1.1.1

# Optional: For GPU acceleration (uncomment if you have CUDA)
# onnxruntime-gpu==1.22.1

# Development and Testing
pytest==7.4.3
requests==2.32.5
